Identfying isotopologue peaks in MS data

Given a spectrum (i.e. a peak matrix with m/z and intensity values) the function identifies groups of peaks that correspond to isotopes

.isotope_peaks(
  x,
  substDefinition = isotopicSubstitutionMatrix(),
  tolerance = 0,
  ppm = 20,
  seedMz = numeric(),
  charge = 1
)

Arguments

x
matrix with spectrum data (columns mz and intensity).

substDefinition
matrix with isotopes definition (columns name, degree, md, min_slope, max_slope). This matrix has to have rows ordered in such a way that the column md is sorted in increasing order.

tolerance
numeric representing the tolerance for the relaxed matching of m/z values of peaks

ppm
numeric(1) representing a relative, value-specific parts-per-million (PPM) tolerance for the relaxed matching of m/z values of peaks

seedMz
numeric ordered vector containing m/z values. If provided, the function checks if there are peaks in x whose m/z match them. If so, it looks for isotope groups related to this subset of peaks.

charge
numeric(1) representing the charge of the ionized compounds

Value

list of vectors. Each vector in the returned list contains the indexes of the rows in x that match a certain isotope group found by the function.

Details

The function iterates over the peaks (rows) in x. Firstly, it checks the presence of peaks whose m/z difference with respect to the current peak matches (the matching can be relaxed via the parameters ppm and tolerance) certain m/z differences related to mass differences of certain substitutions in substDefinition and the chosen charge. Then, if any such peak is found, the function checks if also their intensity is compatible with them being part of a isotopic group and if so they are grouped together. When some peaks are grouped together their indexes are excluded from the set of indexes that are searched for further groups.

Examples